Get Organization Values

Input basic information about an organization and return a list of more detailed information.

Usage

get_org_values(
  state = "CA",
  location.type = "rural",
  total.expense = 1e+06,
  ntee = "A20"
)

Format

get_org_values()

Arguments

state

two letter abbreviation of which state your organization is located in. See state.abb52 for options.

location.type

either "rural", "town", "subruban" or "metro" for the type of city your organization is located in.

total.expense

a number of the total annual expenses of your organization in U.S. dollars.

ntee

a character string of the NTEE code your organization belongs to. A list of all available NTEE codes can be found in the ntee.orig data set and at https://nccs.urban.org/publication/irs-activity-codes.

Value

A list of characteristics about your organization

Examples

get_org_values(state = "NY", location.type = "rural", total.expense = 5000000, ntee = "K93")
#> $total.expense
#> [1] 5e+06
#> 
#> $state
#> [1] "NY"
#> 
#> $location.type
#> [1] "rural"
#> 
#> $old.code
#> [1] "K93"
#> 
#> $new.code
#> [1] "RG-HMS-K93"
#> 
#> $type.org
#> [1] "RG"
#> 
#> $broad.category
#> [1] "HMS"
#> 
#> $major.group
#> [1] "K"
#> 
#> $univ
#> [1] FALSE
#> 
#> $hosp
#> [1] FALSE
#> 
#> $two.digit
#> [1] "93"
#> 
#> $further.category
#> [1] NA
#> 
#> $division.subdivision
#> [1] "93"
#> 
#> $division
#> [1] "9"
#> 
#> $subdivision
#> [1] "3"
#> 
get_org_values(state = "DC", location.type = "metro", total.expense = 210000, ntee = "A0161")
#> $total.expense
#> [1] 210000
#> 
#> $state
#> [1] "DC"
#> 
#> $location.type
#> [1] "metro"
#> 
#> $old.code
#> [1] "A0161"
#> 
#> $new.code
#> [1] "AA-ART-A61"
#> 
#> $type.org
#> [1] "AA"
#> 
#> $broad.category
#> [1] "ART"
#> 
#> $major.group
#> [1] "A"
#> 
#> $univ
#> [1] FALSE
#> 
#> $hosp
#> [1] FALSE
#> 
#> $two.digit
#> [1] "01"
#> 
#> $further.category
#> [1] 61
#> 
#> $division.subdivision
#> [1] "61"
#> 
#> $division
#> [1] "6"
#> 
#> $subdivision
#> [1] "1"
#>