nonprofits

All nonprofits that electronically filled with the IRS from 2009-2019. Each nonprofit is listed in this data set exactly once at their most recent file time. See vignette for detailed description on how this data set was created.

This is all non profits available for use when generating an appraisal.

Use EIN to match with nonprofits.detailed data set for more detailed information about each nonprofit.

Usage

nonprofits

Format

nonprofits

A data frame with 12,847 rows and 18 columns. Each row is a unique nonprofit, and the columns are its respective attributes:

EIN

nonprofit IRS employer identification number

form.year

IRS Filing year

name

Name of nonprofit

url

URL of ProPublica.org with more detailed information about the nonprofit.

ntee

Original NTEE code of nonprofit. See references for details

new.code

New disaggregated NTEE code. See vignette for details.

univ

TRUE or FALSE, is the nonprofit a university?

hosp

TRUE or FALSE, is the nonprofit a hospital?

total.expense, total.employee, gross.reciepts, total.assests

respective items reported to the IRS for the filing year

ceo.compensation

Total CEO compensation

gender

Imputed gender of the CEO, see references for details.

state

State the nonprofit is located in

zip5

5 didget zip code the nonprofit is located in

location.type

"metro", "subruban", "town", or "rural", what type of city the nonprofit is located in